Re-order files to stop polluting generic files
