+++ lua/keymap/keymapper.lua
@@ -207,10 +207,18 @@ function GetKeyActions()
     local debugKeyActions = import("/lua/keymap/debugkeyactions.lua").debugKeyActions
 
     for k,v in keyActions do
+        if ret[k] and ret[k] != v.action then
+            WARN(string.format("Overwriting user key action: %s -> %s", k, ret[k].action))
+        end
+
         ret[k] = v
     end
 
     for k,v in debugKeyActions do
+        if ret[k] and ret[k] != v.action then
+            WARN(string.format("Overwriting user key action: %s -> %s", k, ret[k].action))
+        end
+
         ret[k] = v
     end
