+++ lua/sim/commands/distribute-queue.lua
@@ -52,7 +52,7 @@ DistributeOrders = function(units, target, clearCommands, doPrint)
         return
     end
 
-    if not target then
+    if (not target) or (IsDestroyed(target)) then
         return
     end
